pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "banana-peel"
version = "0.1.0"
description = "Remove Gemini watermarks and losslessly compress PNG files"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["gemini", "watermark", "png", "compression", "image"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.20.0",
    "pyoxipng>=9.0.0",
    "watchdog>=3.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.scripts]
banana-peel = "banana_peel.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/kendaniels/banana-peel"
Source = "https://github.com/kendaniels/banana-peel"
Issues = "https://github.com/kendaniels/banana-peel/issues"
Changelog = "https://github.com/kendaniels/banana-peel/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
banana_peel = ["assets/*.png"]
